requirements.txt
asgiref==3.8.1
certifi==2024.7.4
Django==5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.9.0
simplejwt==2.0.1
sqlparse==0.5.1
typing==3.7.4.3
